+++ eras/babbage/impl/src/Cardano/Ledger/Babbage/UTxO.hs
@@ -47,8 +47,7 @@ instance EraUTxO BabbageEra where
 
   getConsumedValue = getConsumedMaryValue
 
-  getProducedValue pp isRegPoolId txBody =
-    withTopTxLevelOnly txBody (getProducedMaryValue pp isRegPoolId)
+  getProducedValue = getProducedMaryValue
 
   getScriptsProvided = getBabbageScriptsProvided